add test cases for review controller

+++ nextjs-app/src/lib/controllers/ReviewController.js
@@ -7,6 +7,8 @@ export default class ReviewController {
   }
 
   async onePathAllReviews(bikepathId) {
+    if (!Number.isInteger(bikepathId) || bikepathId <= 0)
+      throw new ValidationError("bikepathId must be a positive integer");
     return await this.dao.onePathAllReviews(bikepathId);
   }
 
@@ -18,30 +20,46 @@ export default class ReviewController {
     const limit_parsed = limit_raw ? parseInt(limit_raw) : null;
     const userId_parsed = userId_raw ? parseInt(userId_raw) : null;
 
+    // Validate parsed values
+    if (!Number.isInteger(limit_parsed) || limit_parsed < 0)
+      throw new ValidationError("Input valid limit and valid userId");
+
+    if (!Number.isInteger(userId_parsed) || userId_parsed < 0)
+      throw new ValidationError("Input valid limit and valid userId");
+
     return await this.dao.allPathsAllReviews(limit_parsed, userId_parsed);
   }
 
   async getStats(bikepathId) {
+    if (!Number.isInteger(bikepathId) || bikepathId <= 0)
+      throw new ValidationError("bikepathId must be a positive integer");
     return await this.dao.onePathStats(bikepathId);
   }
 
   async createReview(score, comment, userId, bikepathId) {
+    if (!Number(score) || score < 0 || score > 5)
+      throw new ValidationError("Invalid score");
+    if (!Number.isInteger(bikepathId) || bikepathId <= 0)
+      throw new ValidationError("bikepathId must be a positive integer");
+    if (!Number.isInteger(userId) || userId <= 0)
+      throw new ValidationError("userId must be a positive integer");
+
     return await this.dao.createNewReview(score, comment, userId, bikepathId);
   }
 
   async updateReview(reviewId, review) {
-    // TO DO: validate rating, userId, etc.
-
+    if (!Number(review.score) || review.score < 0 || review.score > 5)
+      throw new ValidationError("Invalid score");
+    if (!Number.isInteger(reviewId) || reviewId <= 0)
+      throw new ValidationError("reviewId must be a positive integer");
     return await this.dao.updateReview(reviewId, review);
   }
 
   async deleteReview(userId, reviewId) {
-
     const review = await this.dao.onePathOneReview(reviewId);
-    console.log(review);
+    if (review.length === 0) throw new ValidationError("Review not found.");
     if (userId !== review.userId) {
-      console.log("not the author uh oh");
-      throw { status: 403, error: "Only the author can delete this review" };
+      throw new Error("Only the author can delete this review");
     }
 
     return await this.dao.deleteReview(reviewId);
